refactored to have Backend trait and two implementations in separate sub-modules
